Creation of only one EditProfile activity against two

+++ app/src/main/java/com/intac/Registration_page2.kt
@@ -53,7 +53,7 @@ class Registration_page2 : AppCompatActivity() {
                 val SignUpResponse = SingUp(user.login, user.pass)
                 println(SignUpResponse.state)
                 val intent = Intent(this@Registration_page2, Feed::class.java)
-                intent.putExtra("id", SignUpResponse.id)
+                intent.putExtra("id", SignUpResponse.id.toLong())
                 startActivity(intent)
             } else {
                 Log.d("TestReg", SignInResponse.state) // в будущем будет соответствующая ошибка на экране
